// src/database.types.ts
export type Database = {
  public: {
    Tables: {
      challenges: {
        Row: {
          created_at: string
          id: string
          user_id: string | null
          value: string
        }
        Insert: {
          created_at?: string
          id?: string
          user_id?: string | null
          value: string
        }
        Update: {
          created_at?: string
          id?: string
          user_id?: string | null
          value?: string
        }
        Relationships: []
      }
      credentials: {
        Row: {
          aaguid: string
          backup_state: Database["public"]["Enums"]["backup_state"]
          created_at: string
          credential_id: string
          credential_type: Database["public"]["Enums"]["credential_type"]
          device_type: Database["public"]["Enums"]["device_type"]
          friendly_name: string | null
          id: string
          last_used_at: string | null
          public_key: string
          sign_count: number
          transports: string[]
          updated_at: string
          user_id: string
          user_verification_status: Database["public"]["Enums"]["user_verification_status"]
        }
        Insert: {
          aaguid?: string
          backup_state: Database["public"]["Enums"]["backup_state"]
          created_at?: string
          credential_id: string
          credential_type: Database["public"]["Enums"]["credential_type"]
          device_type: Database["public"]["Enums"]["device_type"]
          friendly_name?: string | null
          id?: string
          last_used_at?: string | null
          public_key: string
          sign_count: number
          transports: string[]
          updated_at?: string
          user_id?: string
          user_verification_status: Database["public"]["Enums"]["user_verification_status"]
        }
        Update: {
          aaguid?: string
          backup_state?: Database["public"]["Enums"]["backup_state"]
          created_at?: string
          credential_id?: string
          credential_type?: Database["public"]["Enums"]["credential_type"]
          device_type?: Database["public"]["Enums"]["device_type"]
          friendly_name?: string | null
          id?: string
          last_used_at?: string | null
          public_key?: string
          sign_count?: number
          transports?: string[]
          updated_at?: string
          user_id?: string
          user_verification_status?: Database["public"]["Enums"]["user_verification_status"]
        }
        Relationships: []
      }
      users: {
        Row: {
          created_at: string
          email: string
          id: string
          name: string | null
        }
        Insert: {
          created_at?: string
          email: string
          id?: string
          name?: string | null
        }
        Update: {
          created_at?: string
          email?: string
          id?: string
          name?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      backup_state: "not_backed_up" | "backed_up"
      credential_type: "public-key"
      device_type: "single_device" | "multi_device"
      user_verification_status: "unverified" | "verified"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof PublicSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof PublicSchema["CompositeTypes"]
    ? PublicSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never